pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "health-economic-impact-simulator"
version = "0.1.0"
description = "A tool for simulating and analyzing health interventions' economic impacts"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["health", "economics", "simulation"]
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "streamlit>=1.29.0",
    "pyyaml>=6.0.1",
    "pandas>=2.1.4",
    "plotly>=5.18.0",
    "numpy>=1.26.2"
]

[project.urls]
Documentation = "https://github.com/username/health-economic-impact-simulator#readme"
Source = "https://github.com/username/health-economic-impact-simulator"

[tool.hatch.build.targets.wheel]
packages = ["src/heis"] 